Update error handling when components are not wrapped by SplitFactoryProvider
