+++ src/constants.ts
@@ -16,6 +16,6 @@ export const CONTROL_WITH_CONFIG: SplitIO.TreatmentWithConfig = {
 // Warning and error messages
 export const WARN_SF_CONFIG_AND_FACTORY: string = '[WARN]  Both a config and factory props were provided to SplitFactoryProvider. Config prop will be ignored.';
 
-export const EXCEPTION_NO_REACT_OR_CREATECONTEXT: string = 'React library is not available or its version is not supported. Check that it is properly installed or imported. Split SDK requires version 16.3.0+ of React.';
+export const EXCEPTION_NO_SFP: string = 'No SplitContext was set. Please ensure the component is wrapped in a SplitFactoryProvider.';
 
 export const WARN_NAMES_AND_FLAGSETS: string = '[WARN]  Both names and flagSets properties were provided. flagSets will be ignored.';
